k/configs: revert broken fix for cleanup.policy
The first attempt at fixing the way to report cleanup.policy was
incorrect, because it meant that we reported the config as
DEFAULT_CONFIG even if the config was explicitly set but to the
cluster-default value.

Instead, in this case, we want to report this as a DYNAMIC_TOPIC_CONFIG.
We can do this now since we stopped setting the cleanup.policy value
explicitly for each topic.

Furthermore, the compat tests was simplified because now there are 2
separate fix attepts deployed at multiple redpanda versions, so to keep
the test sane, we just ignore the diff until a version (24.2.1) where we
know that this version will contain the fixes and all following version
we might compare against also contains the fixes.

Lastly, the change to the DescribeTopicConfigs ducktape test was also
reverted to correctly report the cleanup.policy as DYNAMIC_TOPIC_CONFIG
when this was explicitly specified on the topic creation request.