storage: Add chained_batch_consumer utility
The utility is supposed to simplify the code that uses fanout streams.
The problem with fanout stream utility is that the progress of one
consumer is tied to the progress of another. One slow or blocked
consumer could potentially block all others. So the code should take
this into account and run things concurrently. This makes the code more
complex.

To avoid this complexity the chain consumer is introduced. The consumer
can accept several consumers and feed them batches from the underlying
stream. It can deal with situation when one consumer accepts batches and
another skips or one of the consumers stops early.