cst: Change acceptable query miss condition
When a query for an offset misses, it is possible that this happens
because the manifest was entirely truncated. Truncation usually happens
in two steps, first the start offset is advanced and later on the
manifest segments are removed.

If the query falls between these two steps, the current condition to
accept a query miss as a non-error may fail, because it uses the kafka
log start offset. If the manifest start offset has been advanced beyond
the entire manifest span, the kafka offset is now not calculatable,
which results in nullopt kafka start, which does not match the condtion
and the query miss is recorded as an error.

The change detects the condition where the manifest is fully truncated
and the query offset lies within truncated range, if so the query is
considered not to have failed.