cst/ducktape: Accept errors due to gap in manifest
The scrubber test may remove the first segment in the manifest. This
causes a difference between the manifest start offset and the first
segment start offset. The mismatch causes a validation to fail when
spillover is applied (which is correct behavior).

To avoid the test failing due to this expected set of events, the error
log is accepted in the test.

(cherry picked from commit 7a9e604111005d3085ac9bfaa91728745021373e)