storage: change map type for `_db` in `kvstore`
In an effort to move to fragmented data structures, the `absl::btree_map`
type used in the `kvstore` is now replaced by a `chunked_hash_map`.

We do not rely on the ordering or any sort of pointer/reference/iterator
stability provided by `absl::btree_map` in our current uses of `_db`, so
this is a painless swap.