config/datalake: add options for datalake topics
This change adds two new configuration options for controlling datalake
support as part of the Iceberg project. One option, at the cluster
level, can be used to globally enable or disable datalake support. The
topic option turns on datalake support for a specific topic. A topic is
considered "datalake-enabled" when both options are set.

This change also adds logging in ntp_archiver_service to indicate the
state of topics as they are uploaded.