mvlog: log basics for segment rolls
Introduces the beginnings of a log implementation that is able to roll
new segments with segment.bytes or segment.ms.

The log, similar to the disk_log_impl, is backed by a circular buffer of
segments, with the noteable difference that the active segment is
managed explicitly and separately from the others.

I'm not fully sold on this approach (e.g. vs only managing the segment
appender separately but keeping the underlying readable segment with the
rest of the segments), but for now it makes reasoning about segment
rolling concurrency a bit more straightforward, and down the line it
will make it easier to reason about segment removals (e.g. the active
segment cannot be removed).

This implementation is still missing quite a bit (e.g. it doesn't flush,
doesn't implement the exact log interface), but at least begins to
introduce some ideas that can back the new log implementation.