rm_stm: redo relink_producer logic
relink_producer was intended to move the producer to the back of the LRU
cache. This was done at different times on the leader/followers. On
leaders, it was done right after the request was processed whereas on
followers it was done after the batch was applied. In order not to do it
twice on leader, the relink_producer was added to detect if it was
already done.

This commit consolidates the logic to do it in apply for all the
replicas and adds a guardrail in can_evict() to prevent eviction of
producers with unapplied batches.