Prevent Rust transform death spiral by not panicking on transform Err.
If a Rust transform returns Err, the runtime panics because it currently
calls expect on the Result. This makes any Err returned by a users
transform put the wasm runtime into a death spiral of panic and restart
loops. Until the runtime has a concept of handling "bad" records that
cause errors (e.g. DLQ, drop, etc.) the transform should be responsible
for this. Returning Err should just result in a log message, at worst.