mvlog/tests: clean up files in test body
I noticed that death tests were leaving behind some directories, even
though the test appears to be exiting successfully.

It turns out that because death tests run a separate process[1], it was
creating an extra directory, but not running the teardown because of the
process crash.

So this hacks the tests to clean up in the test body before the crash.

[1] https://github.com/google/googletest/blob/main/docs/advanced.md#death-tests-and-threads