rptest: fix retention value in `archive_retention_test`
The `retention_value` for `retention.bytes` in
`archive_retention_test.test_delete` was being set to a value of `1KiB`
by default. Sometimes, a segment smaller than `1KiB` would be produced,
and thus would not be garbage collected.

The `retention_value` is now set using the minimum segment size to ensure
all segments are properly deleted per the expectations of the test.