cloud_storage_clients/configuration: fips_mode in make_configuration
cloud_storage::configuration::get_s3_config:
read fip mode from node config to pass it to
s3_configuration::make_configuration.

this bool is used to set virtual_host mode and skip
s3_client::self_configure

fips mode in s3 supports only url_style==virtual_host, and fips mode
requires that the endpoints to be properly formatted in a virtual_host from the start, so if url_style
is left without a value, assume that it is virtual_host, and run
self_configuration to check that TS is correctly configured.

if url_style is set to path, assert and terminate.