cloud_storage_clients/configuration: construct virtual_host endpoint_uri
this is critical for fips mode: server_addr needs to point to the full
virtual_host uri (if using virtual_host style, a requirement for fips)

to do so, wire in bucket_name in make_configuration.

note: s3_client will build its full uri for the HTTP Host field, to
support self discovery (not in fips mode). so it still get only the base
s3 endpoint as input