producer_state: add transaction state tracking
Adds the core state tracking (not hooked yet) to the producer state.
Moving forward this is the only place to lookup open transaction state.

This state encapuslates all the in progress transaction state for a
given producer to a transaction. The state machine maintains a list of
active producers to the partition (coming in the next commits).