producer_state: make eviction synchronous
Prior to this commit producer_state::evict() is asynchronous because it
waited on a gate to drain out any pending tasks on the producer state.
This resulted in a fiber per evict() call in the producer_state_manager
resulting in large number of fibers when evicting a ton of producers in
one eviction tick (which manifested as a large allocation).

This commit fixes the issue by making evict() synchronous and getting
rid of the gate that was forcing it. That is ok because the eviction
tick runs synchronously (without scheduling points) and only considers
candidates that are still linked to the shard wide list. Any caller
function using the producer state either waits on op_lock or detaches
from the list before running the operation ensuring that it is not
considered for eviction.