tests: fixed recently added balancer simulator test case
The test case that was recently added to simulator based partition
balancer tests was flaky as we did not wait for enough balancer rounds
for all replicas to be moved. Modified the test to wait for the number
of rounds which is greater than 3 times the number of replicas. This
will make the test stable as we will allow balancer to execute.

Signed-off-by: Michał Maślanka <michal@redpanda.com>