storage: Make log_retention_tests use falloc step size as leeway
This is in preparation for also `truncate`ing segment files which will
adapt their logical size on disk.

Note that this test is a bit different to how prod works because in
reality segments get closed and truncated to their real size on roll so
only ever the last segment has unwritten blocks but this test doesn't
work like that and hence we need account for the unwritten blocks in all
segments.