r/stm_manager: do not allow reading beyond next offset
In state machine manager the background apply fiber is there to apply
batches to state machines which are behind the main apply fiber. The
background apply fiber is only active if an stm is behind what is
currently being applied by the main fiber. A bug in reader configuration
lead to a situation in which a background apply could read up to `_next`
offset of the state machine manager leading to the situation in which a
stm `last_applied_offset` is equal to `_next` therefore the stm next
offset is actually equal to `state_machine_manager::_next + 1`. This
lead to a situation in which there may be no stms suitable to be
handled in main apply fiber but also they are not recoverable by
background apply fiber as their `_next` offset is greater of the one
from the state machine manager. This causes a stall in the
`state_machine_manager` and state machines stop making progress.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 0f8963c9b3a006d4bf0edb0bb46b5414935589eb)