remote_scripts: Make compute_storage.py actually read segments
We used compute_storage.py for validation checks to confirm segments
"are the same" after certain operations. It uses the logical file size
for that.

With the switch to the segment_appender to also adjust the logical file
size when we fallocate this would break.

Make compute_storage actually read the segment files and parse the
actual amount of batch data in there.

We apply a small optimization where for fully closed/truncated segments
we still use the file size and skip reading the full file.

We also switch `data_checksum` to use compute_storage.py and calculate
the md5 sum in compute_storage.py. This avoids repeated invocations of
python which can slow things down.