r/consensus: handle forced configuration changes without appends
Changed the way how Redpanda raft handles force reconfigurations.
Instead of appending configuration to the log we keep a volatile in
memory override for the configuration that was forced. The override is
dropped whenever a node receives different configuration. This approach
allow us to avoid discrepancies as not data are appended to the log
without the active leader.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 01bef11f442270dc632f39a19ee246581f4a9ca8)