cloud_storage/configuration: get_s3_config url_style_override
pass a value that will take precedence over cluster property
"url_style".

default to nullopt.

this is to support auto discovery what style of url the s3 backend will
use:

if "url_style" is set for autodiscovery, a s3 client will be
consstructed, discover the value for url_style, and then a new
configuration will be built passing an actual value