test: Fix test_full_disk_triggers_gc
In the test we trigger a segment GC and expect all data to be
effectively gone.

This is still the case but because the check was looking at the whole RP
data dir it would now pick up the controller and kvstore log files which
also have their logical file size adapted.

Only look at the kafka dir.