rpk: Increase default controller log size limit
Otherwise the controller log would now be skipped because on the average
system where the default falloc step is 32MB it would now already be too
big for the default 20MB limit.

Note that this doesn't actually bloat the size of the debug bundle as
any zero padding basically compresses perfectly anyway.

20 MB also feels kinda low for something important as the controller
log.