security: create config-specific header
The security_config library contains specific bits that are shared
between security module and configuration. However, the header
declarations for those bits are in headers that aren't easily included
in configuration system because of all the additional things brought
along.

So this commit splits these out into a specific config.h header that is
intended to be included by configuration system.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>