config: remove bottomless bucket max width
It's nice that the bottomless bucket exposes a max width that can be
used in configuration system, but its not scalable to do this for
everything in the tree otherwise the configuration subsystem is going to
have too many dependencies.

In this case the max width seems to be about type compatibility in that
its not clear that choosing max int would ever be a reasonable choice
anyway. So external configuration facing validation is just different
and can live in config system.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>