rpk: allow seeking a non-existent group
It should be possible to set commits for a group that does not yet
exist. This was not previously possible because the first thing rpk did
when seeking was to query old offsets. If old offsets did not exist, rpk
would exit with an error that offsets were unable to be fetched.