tx/tests: fix check_sequence_table_cleaning_after_eviction_test
This test used reduction in memory usage
(tx_mem_tracker_consumption_bytes) as a proxy for reduction in number of
producers loaded from the snapshot. This metric was removed as we no
longer (need to) track the bytes used by the strutures. This commit
rewrites the validation using describe producers API.