tests: adapt partition move tests to node-local core assignment
In tests we need to support both the old way to dispatch x-core
movements (via admin.set_partition_replicas) and the new way (via
admin.set_partition_replica_core) depending on whether the cluster is
new or in the process of upgrade. To allow that, change functions in
PartitionMovementMixin:
1) allow omitting "core" field in replica assignment dicts (absent field
   means that the core doesn't matter)
2) add node_local_core_assignment flag to switch between the old and the
   new way.

Also, modify available_policy for the feature flag to ensure that for
most tests node-local core assignment will be enabled.