rptest: use large keyspace in swarm
client-swarm default to "1000" for the --keys argument if you don't
provide it, which sets the maximum number of unique keys that will be
used when producing messages. This means that at most 1000 partitions
(and in practice quite a bit fewer, due to collisions) can be
produced to which is quite unexpected and leads to imbalanced load.

Instead, default to "unlimited keyspace" (largest prime less than
2^63).