kafka/quota_mgr: introduce tracker_key key type
This commit is intended to be a refactor, but the only behaviour change
we expect is for the edge case when a client quota group name clashed
with a client id in the map when represented as a sstring. Previously
such name clashes could happen (although this was not intentional), now
these clashes cannot happen because the strong typing prevents this as
`k_client_id{"abc"} != k_group_name{"abc"}`.