schema_registry: Fix for adding previously deleted schema
store::get_versions was ignoring include_deleted, which meant that
the new schema was detected as existing, but not a candidate for
being added as a new version.

A test was fixed up, as schema id 1 is not compatible with id v2,
compatibility must be set to NONE before adding back id 1, as it
will become version 3.

Signed-off-by: Ben Pope <ben@redpanda.com>