cloud_storage: Add cloud storage cache flattening
The names that start with randomized prefix are now a subject to
rehashing procedure. The prefix is replaced with another prefix which is
computed from the hash. This new prefix has lower cardinality.

During the lookup we will try to probe two paths. The rehashed and
original because the cache might store objects from before this feature
was enabled.