k/quota_translator: fix quota values for group-based keys
This fixes a bug in the quota values being returned for group-based
quota keys where we would incorrectly fall back to the default values
for group-based keys when the correct quota is really "not applicable"
(i.e. `std::nullopt`).

This is not an observable bug because these quota values are never
really accessed. If the request had a group-based applicable quota, it
would both have a group-based key and its limit. It's not possible for a
request to have a group-based key and a default quota value.

We fix this bug now because we are about to integrate with the quota
store, which makes this bug more apparent.