k/client_quota_translator_test: simplify group-based test case
The client-id-prefix quotas configured through the Kafka API are going
to have their group names be equal to their client id prefix.

Therefore, to allow us to reuse the existing cluster config-based test,
we simplify it here to have both the produce and fetch configs have
their group names be equal to the client id prefix.