tls_config: Add require_crl param to get_credentials_builder
And introduce stub conditional for switching verification on.

This has no effect with a gnutls backend, where CRL checks are enabled
by default but are also conditional on the presence of a CRL file. If
no CRL file is configured, that is considered a verification PASS, and
there is no way to configure ourselves around this.

However, once we fully migrate to openssl, where CRL checks are disabled
by default, enabling checks carries with it a requirement that a CRL file
is present, and we will use the require_crl param to trigger this regardless
of whether a crl file has been config'd.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>