cst/cache: Sync tracker periodically
The tracker is synced periodically with on-disk data using a directory
walk. This is necessary because once we switch to trimming using the
tracker, the data in tracker and on-disk may drift because of tmp
files, aborted downloads or manual deletion of files. A periodic sync
brings the tracker in sync with the metadata on disk.

(cherry picked from commit ee90a7b52c396b07bdb7dea2dc783020f42ed9a4)