rpk: display float values as float in cluster config get
In 610e3ad8 we introduced a fix for #6070, to avoid
printing large numbers with scientific notation, we
treated all numbers as integers.

Now, the API uses float values and they are being
rounded, this change introduce a mod check to
determine whether the number is an integer or a
float.