kafka: add `read_distribution_probe`
Add a new probe that will be used in the `kafka::server`.
Contains the `log_hist_read_dist` object for recording the
read distribution of data fetched, based on the timestamp delta
between the data read and the current `now()` timestamp.

Histogram metrics is aggregated across shards when cluster config
`aggregate_metrics` is `true`.