rptest: increase initial connections threshold
Before test_max_connections starts the main body of the test we assert
that there are a "small" number of connections.

This number is not zero because various things maintain or make
occasional connections to cloud clusters like kminion, so we need to
choose a higher value.

I observed typically ~45 connections on an idle T1 and 450
on an idle T7 cluster. This led to an initial SWAG of 100 + 1% of the
connection target as the threshold above which we complain.

However, this check failed on a GCP T4 run with 320 connections, versus
250 allowed. So just double the calculation to 2% + 100. This threshold
will automatically become much larger once the connection count
increase makes it into the product data.

PR has internal slack ref.