k/quota_mgr: add metrics on token bucket state
This introduces metrics on the token bucket states inside quota
manager. These metrics are coarse as they are aggregated at the level of
(client_quota_rule, client_quota_type) combinations, but they should be
useful for an overview of what kinds of quotas have been configured and
apply to clients.