dt/quotas: refactor producer/consumer creation
There are many places in this file where producers and consumers are
created with the same arguments. This refactors the tests to create the
producers and consumers using factory methods.

This also sets retries=2 and request_timeout_ms=60000 consistently on
all producers. An earlier commit set this for one of the tests
referencing CI failures, so it makes sense to be defensive with this
config and set this more widely.