container: interval set mutable end
std::btree_set in newer abseil returns const iterators (like std::set -
https://en.cppreference.com/w/cpp/container/set/begin). the rationale
for this is that the container is indexing a T but the compiler can't
determine which parts of T are used in the comparison function, so it
stays conservative.

i manually looked and it seemed as tho the end field was never involved
in comparisons, so this is probably safe. but btree_map would probably
be clearer where you put the mutable parts in the value.

we could also remove and then re-insert to make compiler happy.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>