tests: adjusted topic settings to make learner gap test stable
The test using decommission operation to validate the correctness of
learner gap metrics used random properties for topic partition count and
replication. This made the test flaky as in some circumstances the
`redpanda/controller/0` was the only partition on the decommissioned
node. This partition is removed from the node and not moved anywhere
else hence it does not contribute to learner gap but it still
contributes to overall data on the decommissioned node.

Fixes: #19067

Signed-off-by: Michał Maślanka <michal@redpanda.com>