rpk: hint error message location for bundle failures
When a command fails to run, rpk will return:

- couldn't save 'foo.txt': exit status 1

And will save stderr in foo.txt for full debugging.
This is not clear, so users may be lost about
what happened and won't know how to get pass this
error. We are adding a hint of where is the rest
of the error (which might be multiple lines of text)