rpk: permission check for k8s bundle
Now we want to check if the authenticated user
account has authorization to collect the k8s
resources needed for the debug bundle process.

If not, we avoid running all the steps and instead
providing a single, meaningful error message
with a hint on how to solve this (link to our docs).

(cherry picked from commit e779bf3bf01aabeca73b543afe12373e4bdf5fea)