cloud_storage: add topic manifest downloader
When performing topic recovery, Redpanda looks for topic manifests for
the target topic, first looking for a binary manifest, and then looking
for a JSON manifest (see cluster/remote_topic_configuration_source.cc).

With the upcoming naming scheme changes, this will need to change to
account for cluster UUID labels.

This commit introduces a topic manifest downloader for this purpose.
A later commit will replace most of `download_topic_manifest()` in
cluster/remote_topic_configuration_source.cc with this new downloader.
For now, this just adds the class and adds some tests.

Note that the new downloader has an optional "hint" that can be used to
differentiate topic manifests from different clusters (or even from
different revisions in the same cluster).