k/group: simplified txs handing in group state machine
Previously transaction state was kept in several map data structures.
It made it harder to understand the transaction flow in consumer group
state machine. Refactored the transaction handling to use a single
source of truth about the transaction state. The group state machine is
now using a `_transactions` map to store the information about all
ongoing transactions. The map is populated when the transaction starts
and the entry in the map is updated throughout the transaction
lifecycle. Finally the entry is removed when transaction resolves and it
is either committed or aborted.

The `group::ongoing_transaction` structure contains all information
about transaction state, pending offset commits and expiration.

Signed-off-by: Michał Maślanka <michal@redpanda.com>