c/p_balancer: fixed possible overflow of final used ration
Replicas eviction policy is asynchronously applied to each partition it
may happen that the size of individual partitions is not exactly equal
to the total usage size reported. In this case the individual partition
sizes that are moved out of the node may sum up to a value larger than
the total used bytes. Added a code preventing overflowing the final
ratio in this case

Signed-off-by: Michał Maślanka <michal@redpanda.com>