xsdk/tests: Update SR integration test to expect named subject
For the golang version:
Add a (useless) ValueSubjectTopicName for the input topic.

Previously, this transform created the schema it needs in the main function,
looking it up by ID and adding to the Serde instance.

The new version moves the subject creation logic to the Serde.Register call,
adding a ValueSubject to produce that effect. This has no functional effect
on the transform and is not something a user would normally do. It is meant
only to confirm that {Key,Value}Subject is working and not interfering with
normal operation.

For the rust version:

Instead of creating subject "demo-topic-value", create "avro-value" so that
we can check the effect of the golang version in integration_test.go. This
also has no functional effect on the test transform.

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>