spill_key_index: Avoid repeated begin() calls
The pattern of calling `absl::*_hash_map::begin()` in a loop is known to
be inefficient. Each time it will have to find the first non-empty
element in the control array.

This results in a O(N^2) complexity because each time we re-walk the
beginning of the array.

Instead keep iterating over the map such that we traverse the control
array only once.

In this case the map will never shrink so we can use the capacity as a
generation counter to avoid concurrent modifaction and iterator
invalidation.