rpk: copy data directory structure to logs in bundle
Our python decoding tool takes the data directory
as input and searches for redpanda/controller/0_0
and starts decoding the files.

Before this change we were saving the folder in a
different structure and the users who wanted to
decode the logs needed to modify the tool or the
directory structure. This is a QOL change that has
no direct impact on UX since the logs can't be read
unless we use our decoding tools.