tests: use partition size metrics in test verifying local retention move
Previously the test was using an on disk size to validate the
correctness of move policy. Switched to use partition size metrics as
they are more accurate and it is easier to assert the correct result.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit d8e8d2d7664bce91e1886fd952612435efd7b7ac)