cloud_storage: introduce remote label
Introduces the remote_label struct, which will be plugged into the topic
properties in a later commit. This label includes a cluster UUID, meant
to indicate the cluster UUID of the cluster that originally created the
topic.

In the future, this label struct can be extended with a user-provided
tag, but for now it will serve as an effective means to avoid collisions
between objects from different clusters.