cloud_storage: add upload_manifest with path arg
I have some upcoming changes that move ownership of deciding a path to a
centralized location. To that end, the cloud_storage::remote will have
less ssay in the path names of manifests.

This commit is a small foundational change to expose an upload_manifest
call in which callers can determine the path. Later changes will begin
using this more extensively, and eventually the call where remote picks
a name will be removed.