cst/inv: Add inventory report parser
A report parser is added which will accept an input stream created
during a report download, and fetch lines from it and provide them to a
consumer. The report in question may optionally be gzip compressed.

The parser is expected to read upto 128 KiB chunks off the stream, one
at a time. In case of a compressed report the stream is decompressed
using 128 KiB chunks as input.

The input report is expected to be a series of rows delimited by
newlines. The inventory report is in the CSV format but the parser is
format agnostic other than expecting newlines to delimit rows in the
input.

There is no upper bound on the size of a single line, so the parser
should not be used where a single input row is very large.