cst/inv: Add an inventory report consumer
The inventory report consumer takes an input stream (or potentially
multiple streams one after another) created from a report download,
extracts rows from the stream using the report parser, and then performs
the following actions per path extracted from the row:

* If the path does not correspond to one of the NTPs led by one of the
  shards on this node, ignore the row.
* Convert the path to a 64 bit hash.
* Append the hash to a vector in-memory for the NTP.
* If the size of total such hashes exceeds a limit, write the vectors to
  disk, to files specific to each of their NTPs, until the size falls
  below a limit.
* Once the entire stream is processed, write all vectors remaining in
  memory to disk, to NTP specific files.

At the end of this processing we have a set of files on disk, one per
NTP, which contain hashes of paths seen for those NTPs inside the
inventory report.

The consumer can be used to consume multiple streams one after another,
the internal map of hashes will be updated across all streams. This is
required for inventories such as GCP where the report is split into
different files if it is above a certain size.

The consumer assumes that the report is within a fixed format,
consisting of "bucket","path". Customizing the format can be done as a
future enhancement.