cst/inv: Add ntp hash loader
Adds a utility which given an NTP, is capable of finding the correct
file written by the inventory report consumer (the parent commit), and
loading it to memory.

Since the inventory report consumer writes out to disk in the form of
vectors of hashes, the loader will have to read the vectors from the
file. It stores the hashes in a set, but before adding a new hash we
check if it is already in the set. If so we are looking at a possible
xxhash collision, because all entries in the inventory report are
supposed to be unique, and therefore all entries in the inventory report
consumer are also supposed to be unique. So an entry already present in
the set and being added again is a hash collision, and is added to a
special collision set.

There is a query method which then allows a caller to check for the
existence of a path. The result is either found, missing or possible
collision.

The scrubber when using this loader should treat possible collisions as
a warning and make a HEAD request for that path.