cluster: use path provider + downloader in partition recovery
The partition recovery manager downloads the partition manifest and
remote segments to be able to create a "deep" copy of the remote replica
that is served by a local Raft log.

This recovery needs to be able to find a partition manifest whose format
(serde or json) is not yet known. This is handled transparently by the
partition manifest downloader, while also taking remote labels into
account.

On top of this, with the naming format of all objects (including
segments) changing, the recovery needs to be able to know what remote
label to use, if any.

To that end, this commit updates the partition recovery path to use a
path provider and partition manifest downloader.

Note, while the source cluster uuid of the path providers will
ultimately be the same, this is a different path provider than what will
be used at runtime by the partition. The latter is owned by the archival
metadata STM and its lifecycle is therefore tied to the partition. The
former is transient and only lives for as long as the recovery is
happening (it is constructed slightly before the partition is created).