ntp_archiver: use partition_manifest_downlaoder for read replicas
Read replicas need to be able to find a partition manifest whose format
(serde or json) is not yet known. This is handled transparently by the
partition manifest downloader, while also taking remote labels into
account.

This commit updates the archiver to use the downloader for the read
replica loop.