r/consensus: refined leadership changed metrics
Previously the leadership changed metric was incremented every time a
group leader was updated with the new leader or the leader was currently
unknown. Technically the change from leader to no leader is a leadership
change but it was confusing to users.

Change the metric updates to only account the changes when leader is
present. This way after each successful leader election the metric is
incremented by one on all partition replicas.

Signed-off-by: Michał Maślanka <michal@redpanda.com>
(cherry picked from commit 5bee838f24b321ba7e61ff783d3250ce1c1914ff)