archival: Add archier_operations_api
Add interface for the archiver oeprations. The idea is to extract
complex multi-step operations out of the ntp_archiver. This will make
the ntp_archiver mostly responsible for invoking the methods of the
object that implements this interface. Currently, all operations are
implemented inside the archiver. This interface mostly covers main data
upload loop and doesn't touch the housekeeping operations.

Another goal of this is to be able to test the archiver independently
from Redpanda (rp_fixture). The archiver_operations_api could be mocked
which will make it possible to test using gmock. It is also possible to
test the archiver_operations_api in isolation using gmock.