archiver: Implement archiver_operations_api
This commit adds incomplete implementation of the interface. It also
defines a set of interfaces used to represent different subsystems in
Redpanda. The list inclues remote api, partition_manager,
cluster::partition, etc. The goal is to be able to mock these interfaces
in thests. They should also be trivial to implement using the actual
Redpanda subsystems and mostly mimic the interfaces these subsystems
have.