archival: fix bug in `upload_should_abort()`
`upload_should_abort()` previously had incorrect logic, in which the
`_parent.term()` was being compared against itself in order to perform
a check for a change in leadership, when the intention was to
compare the `_parent.term()` against `_start_term`.

This commit corrects the logic, and adds `ntp_archiver::lost_leadership()`
to abstract out `_parent` partition leadership checking into its own function.