k/quotas: watch quota store changes
This adds a watch method to `quota_store` to allow watching for updates
done to the store.

It also integrates this with the `quota_manager` to start watching for
changes to the quota store on shard 0. Note that it is safe to watch for
quota changes only on shard 0 because we only update the tracker map and
the tracker limits inside `quota_manager` on shard 0.