cloud_storage: add manifest exists to downloader
Topic recovery validation checks if partition manifests exist, and use a
method that's built into the cloud_storage::remote. In the effort to
change our object naming scheme, this means this needs to be updated to
support the new naming.

Rather than adding more business logic into the remote, this adds a
method to the new partition_manifest_downloader to check if a given
partition's manifest exists.

A later commit will plumb this into topic recovery validation.